refactor(api): update types in Employee, Operation & Unit models

+++ API/Models/Operation.cs
@@ -3,7 +3,7 @@
     public class Operation
     {
         public int OperationID { get; set; }
-        public int? EmployeeID { get; set; }
+        public long? EmployeeID { get; set; }
         public int UnitID { get; set; }
         public int OperationType { get; set; }
         public DateTime IssuedAt { get; set; }
