+++ API/Models/Unit.cs
@@ -3,7 +3,7 @@
     public class Unit
     {
         public int UnitID { get; set; }
-        public int CompanyID { get; set; }
+        public long CompanyID { get; set; }
         public bool IsActive { get; set; }
         public float Lat { get; set; }
         public float Lng { get; set; }
